+++ CHANGELOG.md
@@ -9,6 +9,7 @@ To install or update LODA, please follow the [installation instructions](https:/
 ### Features
 
 * Automatic clean up of local programs (not on Ubuntu 18)
+* New metric for removed programs
 
 ## v21.12.15